sigh

+++ src/cos_models/player_history.rs
@@ -10,7 +10,7 @@ pub mod cos_player_history {
     pub struct PlayerHistory {
         pub log: Vec<Log>,
         pub summary: Vec<Summary>,
-        pub roles_in_clans: Vec<i32>,
+        pub roles_in_clans: Vec<i64>,
         pub clans_map: HashMap<String, ClansMap>,
     }
 
@@ -38,7 +38,7 @@ pub mod cos_player_history {
         pub tag: String,
         pub duration: i64,
         pub count: i32,
-        pub roles: Vec<i32>,
+        pub roles: Vec<i64>,
     }
 
     #[derive(Debug, Serialize, Deserialize)]
